Move code into package object
